Cleanup unused actions
