+++ frontend/src/INDI-Server/INDIDeviceContainer.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux'
 import { INDIDevicePage } from './INDIDevicePage'
-import { indiDeviceContainerSelector } from './selectors-redo';
+import { indiDeviceContainerSelector } from './selectors';
 
 export const INDIDeviceContainer = connect(
   indiDeviceContainerSelector,
