+++ frontend/src/INDI-Server/INDIDeviceGroupContainer.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux'
 import { INDIDeviceGroup } from './INDIDeviceGroup'
-import { indiDeviceGroupSelector } from './selectors-redo';
+import { indiDeviceGroupSelector } from './selectors';
 
 
 export const INDIDeviceGroupContainer = connect(
