+++ frontend/src/INDI-Server/INDIPropertyRowContainer.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux'
-import { indiPropertyRowSelector } from './selectors-redo';
+import { indiPropertyRowSelector } from './selectors';
 import { INDIPropertyRow } from './INDIPropertyRow';
 
 export const INDIPropertyRowContainer = connect(indiPropertyRowSelector)(INDIPropertyRow);
