+++ frontend/src/INDI-Server/INDIServerContainer.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux'
 import INDIServerPage from './INDIServerPage'
-import { indiServerContainerSelector } from './selectors-redo';
+import { indiServerContainerSelector } from './selectors';
 
 const INDIServerContainer = connect(
   indiServerContainerSelector,
